+++ docs/conf.py
@@ -36,7 +36,7 @@ project = "COMPAS"
 copyright = "Block Research Group - ETH Zurich"
 author = "Tom Van Mele"
 
-release = "1.10.0"
+release = "1.11.1"
 version = ".".join(release.split(".")[0:2])
 
 master_doc = "index"