+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Added halfedge loops in `compas.datastructures.Halfedge.halfedge_loop`.
 * Added halfedge strips in `compas.datastructures.Halfedge.halfedge_strip`.
+* Added boundingbox to `compas_rhino.conduits.BaseConduit`
 
 ### Changed